backend/coffee_consumer_simulator/pyproject.toml
[project]
name = "coffee-consumer-simulator"
version = "0.1.0"
description = "Locust-based simulator for coffee order load testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "locust>=2.32.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coffee_consumer_simulator"]
